+++ packages/gasket-helper-intl/test/index.test.js
@@ -28,4 +28,45 @@ describe('LocaleUtils', function () {
       assume(call).throws('Not available in browser');
     });
   });
+
+  describe('.getFallbackLocale', function () {
+    let utils, mockConfig;
+
+    beforeEach(function () {
+      mockConfig = {
+        manifest: require('./fixtures/mock-manifest.json')
+      };
+    });
+
+    afterEach(function () {
+      delete require.cache[require.resolve('./fixtures/mock-manifest.json')];
+    });
+
+    [
+      ['fr-CA', 'fr', 'en-US'],
+      ['fr', 'en-US', 'en-US'],
+      ['fr-CA', 'fr', 'en'],
+      ['fr', 'en', 'en'],
+      ['en-CA', 'en-US', 'en-US'],
+      ['en-CA', 'en', 'en'],
+      ['en-US', 'en', 'en-US'],
+      ['en-US', 'en', 'en'],
+      ['en', null, 'en-US'],
+      ['en', null, 'en']
+    ].forEach(
+      function ([locale, expected, defaultLocale]) {
+        it(
+          `properly falls back to ${expected} when the locale is ${locale} and the default locale is ${defaultLocale}`,
+          function () {
+            mockConfig.manifest.defaultLocale = defaultLocale;
+            utils = new LocaleUtils(mockConfig);
+
+            const actual = utils.getFallbackLocale(locale);
+
+            assume(actual).to.equal(expected);
+          }
+        );
+      }
+    );
+  });
 });